+++ src/components/Card.css
@@ -1,0 +1,53 @@
+.card {
+  display: flex;
+  flex-direction: column;
+  width: 34rem;
+  height: 33.4rem;
+  border-radius: 1.5rem;
+  box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
+  overflow: hidden;
+}
+
+.card-image-container {
+  overflow: hidden;
+}
+
+.card-image {
+  width: 100%;
+  height: 20rem;
+  object-fit: cover;
+  transition: ease-in-out 300ms;
+}
+
+.card:hover .card-image {
+  transform: scale(1.3);
+}
+
+.card-info {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  padding: 1.5rem 2rem;
+}
+
+.card-log-time {
+  color: #666;
+  font-size: 1.3rem;
+}
+
+.card-description {
+  height: 4.9rem;
+  align-self: stretch;
+  overflow: hidden;
+  color: #000;
+  text-overflow: ellipsis;
+  /* white-space: nowrap; */
+  font-size: 1.6rem;
+  line-height: 150%;
+}
+
+.card-date {
+  height: 1.9rem;
+  color: #333;
+  font-size: 1.4rem;
+}
